added Home.js component and moved App.js in
